+++ app/Http/Controllers/Controller.php
@@ -19,6 +19,7 @@ use WNowicki\Generic\Logger\PsrLoggerTrait;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Exceptions\RedirectException;
 use Illuminate\Database\Eloquent\Builder;
+use App\Basket\Merchant;
 
 /**
  * Class Controller
@@ -45,6 +46,20 @@ abstract class Controller extends BaseController
             'info'      => session()->get('info'),
             'error'     => session()->get('error'),
         ];
+    }
+
+    /**
+     * Get Merchant Token
+     *
+     * @author MS
+     * @return string
+     * @throws RedirectException
+     */
+    protected function getMerchantToken()
+    {
+        $merchant = $this
+            ->fetchModelById((new Merchant()), $this->getAuthenticatedUser()->merchant_id, 'merchant', '/');
+        return $merchant->token;
     }
 
     /**